Add data preparation script and update data setup
This commit includes the addition of a new script 'prepare_data.py' in the 'drivisafe' directory. This script is responsible for preparing the data in image frames for further processing.

Additionally, changes have been made to 'data_setup.py'. These changes are not staged for commit yet, and they likely include updates or fixes to the data setup process.